MDFU-10: Validate frame size for a received frame.
In case some bytes are lost during frame reception
e.g. buffer overflow or framing error, we need to
ensure that the size at least meets the minimum
so that we do not have a memory access violation
when we access the CRC.

Disregard any error on MAC layer and continue
discarding incoming wehn looking for the frame
start. This way the receive buffer will be empty
for the next attempt.