MDFU-14: Opt-in for frame logging on transport layer.
Frame logging can take some time depending on the
logger implementation e.g. when it is going directly
through a UART HW module without any buffering.
While the MDFU host is busy logging a command frame
the client could already send back the response
frame on the serial transport (other transports are
not affected due to their synchronous nature), and
when the mac layer is not buffered the data might
get lost.

This patch adds a compiler definition that can
enable logging of transport frames and this is by
default off.